Move screens to newly created folders for organization
