--- src/screens/AppointmentsScreen/AppointmentsScreen.js
@@ -1,14 +1,0 @@
-import { StyleSheet, Text, View } from 'react-native'
-import React from 'react'
-
-const AppointmentsScreen = () => {
-  return (
-    <View>
-      <Text>AppointmentsScreen</Text>
-    </View>
-  )
-}
-
-export default AppointmentsScreen
-
-const styles = StyleSheet.create({})
