+++ src/screens/MainScreens/JournalScreen/JournalScreen.js
