+++ src/screens/ServicesScreens/AppointmentsScreen/AppointmentsScreen.js
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const AppointmentsScreen = () => {
+  return (
+    <View>
+      <Text>AppointmentsScreen</Text>
+    </View>
+  )
+}
+
+export default AppointmentsScreen
+
+const styles = StyleSheet.create({})
