+++ src/screens/ServicesScreens/ChurchServicesScreen/ChurchServicesScreen.js
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const ChurchServicesScreen = () => {
+  return (
+    <View>
+      <Text>ChurchScreen</Text>
+    </View>
+  )
+}
+
+export default ChurchServicesScreen
+
+const styles = StyleSheet.create({})
