+++ src/screens/ServicesScreens/DonationScreen/DonationScreen.js
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const DonationScreen = () => {
+  return (
+    <View>
+      <Text>Hi! This is the donation screen</Text>
+    </View>
+  )
+}
+
+export default DonationScreen
+
+const styles = StyleSheet.create({})
